+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -1812,8 +1812,7 @@ function Assert-M365DSCBlueprint
         # types contained within the BluePrint;
         Write-Host "Initiating the Export of those ($($ResourcesInBluePrint.Length)) components from the tenant..."
         $TempExportName = (New-Guid).ToString() + ".ps1"
-        Export-M365DSCConfiguration -Quiet `
-            -Components $ResourcesInBluePrint `
+        Export-M365DSCConfiguration -Components $ResourcesInBluePrint `
             -Path $env:temp `
             -FileName $TempExportName `
             -Credential $Credentials